refactor: extract proxy entities
